First pass at Evennia_builder
This is a rough and currently not completely working version of a builder for JSON output into Evennia batchcode. The batchcode currently returns a syntax error relating to adding attributes to 'None'. But gives an idea of what it will look like when completed. Currently only creates rooms and exits as objects and mobs do not have locations in the JSON output.