Objects and Mobs now spawn.
Threw another hour into this. The mobs and objects now spawn and are placed in the world.  Resets G and E don't seem to hold enough information. G is to give an object to a mob but only contains the refnum of the object. In any case, walking around the world is a bit more interesting.